requirements.txt
langchain==1.2.15
langchain-text-splitters==1.1.2
langchain-google-genai==4.2.2
langchain-chroma==1.1.0
langchain-community==0.4.1
chromadb>=0.5.0
google-genai==1.73.1
entsoe-py>=0.6.0
feedparser>=6.0.0
streamlit==1.56.0
apscheduler>=3.10.0
pypdf>=4.0.0
requests>=2.31.0
python-dotenv>=1.0.0
pandas>=2.0.0
pydantic>=2.0.0
tqdm>=4.66.0
pytest>=8.0.0
pytest-cov>=5.0.0
# v2 stack
numpy>=1.26.0
scikit-learn>=1.5.0
statsmodels>=0.14.0
xgboost>=2.0.0
pyarrow>=15.0.0
anthropic==0.97.0
